refactor: grpc 연결
