+++ src/gen/com/hearlers/v1/model/counsel_prompt_pb.ts
@@ -10,7 +10,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file com/hearlers/v1/model/counsel_prompt.proto.
  */
 export const file_com_hearlers_v1_model_counsel_prompt: GenFile = /*@__PURE__*/
-  fileDesc("Cipjb20vaGVhcmxlcnMvdjEvbW9kZWwvY291bnNlbF9wcm9tcHQucHJvdG8SFWNvbS5oZWFybGVycy52MS5tb2RlbCJ+CgRUb25lEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSDAoEYm9keRgDIAEoCRISCgpjcmVhdGVkX2F0GAQgASgJEhIKCnVwZGF0ZWRfYXQYBSABKAkSFwoKZGVsZXRlZF9hdBgGIAEoCUgAiAEBQg0KC19kZWxldGVkX2F0IvkBChBDb3Vuc2VsVGVjaG5pcXVlEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSDwoHdG9uZV9pZBgDIAEoCRIPCgdjb250ZXh0GAQgASgJEhMKC2luc3RydWN0aW9uGAUgASgJEiYKGW5leHRfY291bnNlbF90ZWNobmlxdWVfaWQYByABKAlIAIgBARISCgpjcmVhdGVkX2F0GAggASgJEhIKCnVwZGF0ZWRfYXQYCSABKAkSFwoKZGVsZXRlZF9hdBgKIAEoCUgBiAEBQhwKGl9uZXh0X2NvdW5zZWxfdGVjaG5pcXVlX2lkQg0KC19kZWxldGVkX2F0QqcBChljb20uY29tLmhlYXJsZXJzLnYxLm1vZGVsQhJDb3Vuc2VsUHJvbXB0UHJvdG9QAaICBENIVk2qAhVDb20uSGVhcmxlcnMuVjEuTW9kZWzKAhVDb21cSGVhcmxlcnNcVjFcTW9kZWziAiFDb21cSGVhcmxlcnNcVjFcTW9kZWxcR1BCTWV0YWRhdGHqAhhDb206OkhlYXJsZXJzOjpWMTo6TW9kZWxiBnByb3RvMw");
+  fileDesc("Cipjb20vaGVhcmxlcnMvdjEvbW9kZWwvY291bnNlbF9wcm9tcHQucHJvdG8SFWNvbS5oZWFybGVycy52MS5tb2RlbCJ+CgRUb25lEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSDAoEYm9keRgDIAEoCRISCgpjcmVhdGVkX2F0GAQgASgJEhIKCnVwZGF0ZWRfYXQYBSABKAkSFwoKZGVsZXRlZF9hdBgGIAEoCUgAiAEBQg0KC19kZWxldGVkX2F0IpQCChBDb3Vuc2VsVGVjaG5pcXVlEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSDwoHdG9uZV9pZBgDIAEoCRIPCgdjb250ZXh0GAQgASgJEhMKC2luc3RydWN0aW9uGAUgASgJEhkKEW1lc3NhZ2VfdGhyZXNob2xkGAYgASgFEiYKGW5leHRfY291bnNlbF90ZWNobmlxdWVfaWQYByABKAlIAIgBARISCgpjcmVhdGVkX2F0GAggASgJEhIKCnVwZGF0ZWRfYXQYCSABKAkSFwoKZGVsZXRlZF9hdBgKIAEoCUgBiAEBQhwKGl9uZXh0X2NvdW5zZWxfdGVjaG5pcXVlX2lkQg0KC19kZWxldGVkX2F0QqcBChljb20uY29tLmhlYXJsZXJzLnYxLm1vZGVsQhJDb3Vuc2VsUHJvbXB0UHJvdG9QAaICBENIVk2qAhVDb20uSGVhcmxlcnMuVjEuTW9kZWzKAhVDb21cSGVhcmxlcnNcVjFcTW9kZWziAiFDb21cSGVhcmxlcnNcVjFcTW9kZWxcR1BCTWV0YWRhdGHqAhhDb206OkhlYXJsZXJzOjpWMTo6TW9kZWxiBnByb3RvMw");
 
 /**
  * @generated from message com.hearlers.v1.model.Tone
@@ -88,6 +88,11 @@ export type CounselTechnique = Message<"com.hearlers.v1.model.CounselTechnique">
    * @generated from field: string instruction = 5;
    */
   instruction: string;
+
+  /**
+   * @generated from field: int32 message_threshold = 6;
+   */
+  messageThreshold: number;
 
   /**
    * @generated from field: optional string next_counsel_technique_id = 7;
