+++ src/gen/com/hearlers/v1/service/counsel_prompt_pb.ts
@@ -12,7 +12,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file com/hearlers/v1/service/counsel_prompt.proto.
  */
 export const file_com_hearlers_v1_service_counsel_prompt: GenFile = /*@__PURE__*/
-  fileDesc("Cixjb20vaGVhcmxlcnMvdjEvc2VydmljZS9jb3Vuc2VsX3Byb21wdC5wcm90bxIXY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UiLwoRQ3JlYXRlVG9uZVJlcXVlc3QSDAoEbmFtZRgBIAEoCRIMCgRib2R5GAIgASgJIj8KEkNyZWF0ZVRvbmVSZXNwb25zZRIpCgR0b25lGAEgASgLMhsuY29tLmhlYXJsZXJzLnYxLm1vZGVsLlRvbmUiLgoQRmluZFRvbmVzUmVxdWVzdBIRCgRuYW1lGAEgASgJSACIAQFCBwoFX25hbWUiPwoRRmluZFRvbmVzUmVzcG9uc2USKgoFdG9uZXMYASADKAsyGy5jb20uaGVhcmxlcnMudjEubW9kZWwuVG9uZSImChNGaW5kVG9uZUJ5SWRSZXF1ZXN0Eg8KB3RvbmVfaWQYASABKAkiTwoURmluZFRvbmVCeUlkUmVzcG9uc2USLgoEdG9uZRgBIAEoCzIbLmNvbS5oZWFybGVycy52MS5tb2RlbC5Ub25lSACIAQFCBwoFX3RvbmUiXAoRVXBkYXRlVG9uZVJlcXVlc3QSDwoHdG9uZV9pZBgBIAEoCRIRCgRuYW1lGAIgASgJSACIAQESEQoEYm9keRgDIAEoCUgBiAEBQgcKBV9uYW1lQgcKBV9ib2R5Ij8KElVwZGF0ZVRvbmVSZXNwb25zZRIpCgR0b25lGAEgASgLMhsuY29tLmhlYXJsZXJzLnYxLm1vZGVsLlRvbmUiZAodQ3JlYXRlQ291bnNlbFRlY2huaXF1ZVJlcXVlc3QSDAoEbmFtZRgBIAEoCRIPCgd0b25lX2lkGAIgASgJEg8KB2NvbnRleHQYAyABKAkSEwoLaW5zdHJ1Y3Rpb24YBCABKAkiZAoeQ3JlYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEkIKEWNvdW5zZWxfdGVjaG5pcXVlGAEgASgLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWUiXAocRmluZENvdW5zZWxUZWNobmlxdWVzUmVxdWVzdBIRCgRuYW1lGAEgASgJSACIAQESFAoHdG9uZV9pZBgCIAEoCUgBiAEBQgcKBV9uYW1lQgoKCF90b25lX2lkImQKHUZpbmRDb3Vuc2VsVGVjaG5pcXVlc1Jlc3BvbnNlEkMKEmNvdW5zZWxfdGVjaG5pcXVlcxgBIAMoCzInLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsVGVjaG5pcXVlIj8KH0ZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlcXVlc3QSHAoUY291bnNlbF90ZWNobmlxdWVfaWQYASABKAkigQEKIEZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlc3BvbnNlEkcKEWNvdW5zZWxfdGVjaG5pcXVlGAEgASgLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWVIAIgBAUIUChJfY291bnNlbF90ZWNobmlxdWUixwEKHVVwZGF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0EhwKFGNvdW5zZWxfdGVjaG5pcXVlX2lkGAEgASgJEhEKBG5hbWUYAiABKAlIAIgBARIUCgd0b25lX2lkGAMgASgJSAGIAQESFAoHY29udGV4dBgEIAEoCUgCiAEBEhgKC2luc3RydWN0aW9uGAUgASgJSAOIAQFCBwoFX25hbWVCCgoIX3RvbmVfaWRCCgoIX2NvbnRleHRCDgoMX2luc3RydWN0aW9uImQKHlVwZGF0ZUNvdW5zZWxUZWNobmlxdWVSZXNwb25zZRJCChFjb3Vuc2VsX3RlY2huaXF1ZRgBIAEoCzInLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsVGVjaG5pcXVlIkQKI1NhdmVDb3Vuc2VsVGVjaG5pcXVlU2VxdWVuY2VSZXF1ZXN0Eh0KFWNvdW5zZWxfdGVjaG5pcXVlX2lkcxgBIAMoCSJrCiRTYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlUmVzcG9uc2USQwoSY291bnNlbF90ZWNobmlxdWVzGAEgAygLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWUyhgkKFENvdW5zZWxQcm9tcHRTZXJ2aWNlEmUKCkNyZWF0ZVRvbmUSKi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVUb25lUmVxdWVzdBorLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZVRvbmVSZXNwb25zZRJiCglGaW5kVG9uZXMSKS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kVG9uZXNSZXF1ZXN0GiouY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZFRvbmVzUmVzcG9uc2USawoMRmluZFRvbmVCeUlkEiwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZFRvbmVCeUlkUmVxdWVzdBotLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRUb25lQnlJZFJlc3BvbnNlEmUKClVwZGF0ZVRvbmUSKi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVUb25lUmVxdWVzdBorLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZVRvbmVSZXNwb25zZRKJAQoWQ3JlYXRlQ291bnNlbFRlY2huaXF1ZRI2LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0GjcuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEoYBChVGaW5kQ291bnNlbFRlY2huaXF1ZXMSNS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbFRlY2huaXF1ZXNSZXF1ZXN0GjYuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxUZWNobmlxdWVzUmVzcG9uc2USjwEKGEZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZBI4LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlcXVlc3QaOS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbFRlY2huaXF1ZUJ5SWRSZXNwb25zZRKJAQoWVXBkYXRlQ291bnNlbFRlY2huaXF1ZRI2LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0GjcuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuVXBkYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEpsBChxTYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlEjwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuU2F2ZUNvdW5zZWxUZWNobmlxdWVTZXF1ZW5jZVJlcXVlc3QaPS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5TYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlUmVzcG9uc2VCsQEKG2NvbS5jb20uaGVhcmxlcnMudjEuc2VydmljZUISQ291bnNlbFByb21wdFByb3RvUAGiAgRDSFZTqgIXQ29tLkhlYXJsZXJzLlYxLlNlcnZpY2XKAhdDb21cSGVhcmxlcnNcVjFcU2VydmljZeICI0NvbVxIZWFybGVyc1xWMVxTZXJ2aWNlXEdQQk1ldGFkYXRh6gIaQ29tOjpIZWFybGVyczo6VjE6OlNlcnZpY2ViBnByb3RvMw", [file_com_hearlers_v1_model_counsel_prompt]);
+  fileDesc("Cixjb20vaGVhcmxlcnMvdjEvc2VydmljZS9jb3Vuc2VsX3Byb21wdC5wcm90bxIXY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UiLwoRQ3JlYXRlVG9uZVJlcXVlc3QSDAoEbmFtZRgBIAEoCRIMCgRib2R5GAIgASgJIj8KEkNyZWF0ZVRvbmVSZXNwb25zZRIpCgR0b25lGAEgASgLMhsuY29tLmhlYXJsZXJzLnYxLm1vZGVsLlRvbmUiLgoQRmluZFRvbmVzUmVxdWVzdBIRCgRuYW1lGAEgASgJSACIAQFCBwoFX25hbWUiPwoRRmluZFRvbmVzUmVzcG9uc2USKgoFdG9uZXMYASADKAsyGy5jb20uaGVhcmxlcnMudjEubW9kZWwuVG9uZSImChNGaW5kVG9uZUJ5SWRSZXF1ZXN0Eg8KB3RvbmVfaWQYASABKAkiTwoURmluZFRvbmVCeUlkUmVzcG9uc2USLgoEdG9uZRgBIAEoCzIbLmNvbS5oZWFybGVycy52MS5tb2RlbC5Ub25lSACIAQFCBwoFX3RvbmUiXAoRVXBkYXRlVG9uZVJlcXVlc3QSDwoHdG9uZV9pZBgBIAEoCRIRCgRuYW1lGAIgASgJSACIAQESEQoEYm9keRgDIAEoCUgBiAEBQgcKBV9uYW1lQgcKBV9ib2R5Ij8KElVwZGF0ZVRvbmVSZXNwb25zZRIpCgR0b25lGAEgASgLMhsuY29tLmhlYXJsZXJzLnYxLm1vZGVsLlRvbmUifwodQ3JlYXRlQ291bnNlbFRlY2huaXF1ZVJlcXVlc3QSDAoEbmFtZRgBIAEoCRIPCgd0b25lX2lkGAIgASgJEg8KB2NvbnRleHQYAyABKAkSEwoLaW5zdHJ1Y3Rpb24YBCABKAkSGQoRbWVzc2FnZV90aHJlc2hvbGQYBSABKAUiZAoeQ3JlYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEkIKEWNvdW5zZWxfdGVjaG5pcXVlGAEgASgLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWUiXAocRmluZENvdW5zZWxUZWNobmlxdWVzUmVxdWVzdBIRCgRuYW1lGAEgASgJSACIAQESFAoHdG9uZV9pZBgCIAEoCUgBiAEBQgcKBV9uYW1lQgoKCF90b25lX2lkImQKHUZpbmRDb3Vuc2VsVGVjaG5pcXVlc1Jlc3BvbnNlEkMKEmNvdW5zZWxfdGVjaG5pcXVlcxgBIAMoCzInLmNvbS5oZWFybGVycy52MS5tb2RlbC5Db3Vuc2VsVGVjaG5pcXVlIj8KH0ZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlcXVlc3QSHAoUY291bnNlbF90ZWNobmlxdWVfaWQYASABKAkigQEKIEZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlc3BvbnNlEkcKEWNvdW5zZWxfdGVjaG5pcXVlGAEgASgLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWVIAIgBAUIUChJfY291bnNlbF90ZWNobmlxdWUi2wEKHVVwZGF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0EhwKFGNvdW5zZWxfdGVjaG5pcXVlX2lkGAEgASgJEhEKBG5hbWUYAiABKAlIAIgBARIUCgdjb250ZXh0GAMgASgJSAGIAQESGAoLaW5zdHJ1Y3Rpb24YBCABKAlIAogBARIeChFtZXNzYWdlX3RocmVzaG9sZBgFIAEoBUgDiAEBQgcKBV9uYW1lQgoKCF9jb250ZXh0Qg4KDF9pbnN0cnVjdGlvbkIUChJfbWVzc2FnZV90aHJlc2hvbGQiZAoeVXBkYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEkIKEWNvdW5zZWxfdGVjaG5pcXVlGAEgASgLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWUiVAojU2F2ZUNvdW5zZWxUZWNobmlxdWVTZXF1ZW5jZVJlcXVlc3QSDgoGdG9uZUlkGAEgASgJEh0KFWNvdW5zZWxfdGVjaG5pcXVlX2lkcxgCIAMoCSJrCiRTYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlUmVzcG9uc2USQwoSY291bnNlbF90ZWNobmlxdWVzGAEgAygLMicuY29tLmhlYXJsZXJzLnYxLm1vZGVsLkNvdW5zZWxUZWNobmlxdWUyhgkKFENvdW5zZWxQcm9tcHRTZXJ2aWNlEmUKCkNyZWF0ZVRvbmUSKi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5DcmVhdGVUb25lUmVxdWVzdBorLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZVRvbmVSZXNwb25zZRJiCglGaW5kVG9uZXMSKS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kVG9uZXNSZXF1ZXN0GiouY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZFRvbmVzUmVzcG9uc2USawoMRmluZFRvbmVCeUlkEiwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZFRvbmVCeUlkUmVxdWVzdBotLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRUb25lQnlJZFJlc3BvbnNlEmUKClVwZGF0ZVRvbmUSKi5jb20uaGVhcmxlcnMudjEuc2VydmljZS5VcGRhdGVUb25lUmVxdWVzdBorLmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZVRvbmVSZXNwb25zZRKJAQoWQ3JlYXRlQ291bnNlbFRlY2huaXF1ZRI2LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkNyZWF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0GjcuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuQ3JlYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEoYBChVGaW5kQ291bnNlbFRlY2huaXF1ZXMSNS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbFRlY2huaXF1ZXNSZXF1ZXN0GjYuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuRmluZENvdW5zZWxUZWNobmlxdWVzUmVzcG9uc2USjwEKGEZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZBI4LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLkZpbmRDb3Vuc2VsVGVjaG5pcXVlQnlJZFJlcXVlc3QaOS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5GaW5kQ291bnNlbFRlY2huaXF1ZUJ5SWRSZXNwb25zZRKJAQoWVXBkYXRlQ291bnNlbFRlY2huaXF1ZRI2LmNvbS5oZWFybGVycy52MS5zZXJ2aWNlLlVwZGF0ZUNvdW5zZWxUZWNobmlxdWVSZXF1ZXN0GjcuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuVXBkYXRlQ291bnNlbFRlY2huaXF1ZVJlc3BvbnNlEpsBChxTYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlEjwuY29tLmhlYXJsZXJzLnYxLnNlcnZpY2UuU2F2ZUNvdW5zZWxUZWNobmlxdWVTZXF1ZW5jZVJlcXVlc3QaPS5jb20uaGVhcmxlcnMudjEuc2VydmljZS5TYXZlQ291bnNlbFRlY2huaXF1ZVNlcXVlbmNlUmVzcG9uc2VCsQEKG2NvbS5jb20uaGVhcmxlcnMudjEuc2VydmljZUISQ291bnNlbFByb21wdFByb3RvUAGiAgRDSFZTqgIXQ29tLkhlYXJsZXJzLlYxLlNlcnZpY2XKAhdDb21cSGVhcmxlcnNcVjFcU2VydmljZeICI0NvbVxIZWFybGVyc1xWMVxTZXJ2aWNlXEdQQk1ldGFkYXRh6gIaQ29tOjpIZWFybGVyczo6VjE6OlNlcnZpY2ViBnByb3RvMw", [file_com_hearlers_v1_model_counsel_prompt]);
 
 /**
  * Tone
@@ -192,6 +192,11 @@ export type CreateCounselTechniqueRequest = Message<"com.hearlers.v1.service.Cre
    * @generated from field: string instruction = 4;
    */
   instruction: string;
+
+  /**
+   * @generated from field: int32 message_threshold = 5;
+   */
+  messageThreshold: number;
 };
 
 /**
@@ -306,19 +311,19 @@ export type UpdateCounselTechniqueRequest = Message<"com.hearlers.v1.service.Upd
   name?: string;
 
   /**
-   * @generated from field: optional string tone_id = 3;
-   */
-  toneId?: string;
-
-  /**
-   * @generated from field: optional string context = 4;
+   * @generated from field: optional string context = 3;
    */
   context?: string;
 
   /**
-   * @generated from field: optional string instruction = 5;
+   * @generated from field: optional string instruction = 4;
    */
   instruction?: string;
+
+  /**
+   * @generated from field: optional int32 message_threshold = 5;
+   */
+  messageThreshold?: number;
 };
 
 /**
@@ -350,7 +355,12 @@ export const UpdateCounselTechniqueResponseSchema: GenMessage<UpdateCounselTechn
  */
 export type SaveCounselTechniqueSequenceRequest = Message<"com.hearlers.v1.service.SaveCounselTechniqueSequenceRequest"> & {
   /**
-   * @generated from field: repeated string counsel_technique_ids = 1;
+   * @generated from field: string toneId = 1;
+   */
+  toneId: string;
+
+  /**
+   * @generated from field: repeated string counsel_technique_ids = 2;
    */
   counselTechniqueIds: string[];
 };
